Refactored cli commands to their own file
