+++ app.py
@@ -11,6 +11,7 @@ rds = SQLAlchemy(app)
 app.debug = True
 
 from models import *
+import cli
 
 def task2Dict(row):
     di = {
@@ -28,7 +29,6 @@ def task2Dict(row):
     return di
 
 
-
 def taskConverter(o):
     # datetime isn't serializable, so "toString" it
     if isinstance(o, datetime):
@@ -37,53 +37,6 @@ def taskConverter(o):
     if isinstance(o, StatusEnum):
         return o.name
 
-
-@app.cli.command('db_create')
-def db_create():
-    rds.create_all()
-    print('Database created!')
-
-@app.cli.command('db_drop')
-def db_drop():
-    rds.drop_all()
-    print('Database dropped!')
-
-@app.cli.command('db_showtable')
-def db_showtable():
-    print(Task.__tablename__)
-
-@app.cli.command('db_seed')
-def db_seed():
-
-    currenttime = datetime.now()
-    endtime = currenttime + timedelta(seconds=600)
-
-    tsk1 = Task(
-                name='Cook Eggs',
-                description='Cooking eggs for the family',
-                priority=1,
-                status='ACTIVE',
-                starttime = currenttime,
-                currenttime = currenttime,
-                createdtime = currenttime,
-                endtime = endtime
-                )
-
-    tsk2 = Task(
-                name='Boil Water',
-                description='Heating up water on the stove',
-                priority=3,
-                status='PENDING',
-                starttime = currenttime,
-                currenttime = currenttime,
-                createdtime = currenttime,
-                endtime = endtime
-                )
-    
-    rds.session.add(tsk1)
-    rds.session.add(tsk2)
-    rds.session.commit()
-    print('Database seeded')
 
 @app.errorhandler(400)
 def invalid_request(e):
